Add routes direcory
