+++ src/main.js
@@ -2,12 +2,9 @@ import Vue from 'vue';
 import App from './App.vue';
 import './registerServiceWorker';
 
-import JVueRouter from 'j-vue-router';
+
 import store from './store'
-
-import ViewVK from "./views/main/ViewVK";
-import ViewMenu from "./views/ViewMenu";
-import ViewPlayer from "./views/ViewPlayer";
+import routes from './router'
 
 
 
@@ -17,42 +14,9 @@ import ViewPlayer from "./views/ViewPlayer";
 
 
 
-const routes = [
-  {
-    path: '',
-    components:{},
-  },
-  {
-    path: '/vk',
-    components: {
-      'main': ViewVK
-    }
-  },
-  {
-    path: '#menu',
-    components: {
-      'menu': ViewMenu
-    },
-  },
-  {
-    path: '#player',
-    components: {
-      'player': ViewPlayer
-    },
-  }
-];
 
 
 
-
-Vue.use(JVueRouter, {
-  basic:{
-    removeHash: true,
-  },
-  routes,
-  // eslint-disable-next-line no-console
-  // hook: (to, from) => console.log(to, from)
-});
